requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pymongo==4.6.1
motor==3.3.2

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery
celery==5.3.4
flower==2.0.1
celery[redis]==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Networking
httpx==0.26.0
aiohttp==3.9.1

# Content Processing
beautifulsoup4==4.12.3
markdownify==0.11.6
lxml==5.1.0
html5lib==1.1

# AI Integration
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
click==8.1.7
pytz==2023.3.post1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3